Move diagnostic methods to the dedicated module.
